*: Only use server as central.
There's some issue with multiple read/write characteristics on a single
service. And also an issue with struct size that's 24, but strangerly
the issue doesn't show up when the size is 28. Maybe some fragmentation
issue.

So keep the struct that's 28, and split the one that's 24 into two
struct one for read one for write.

It seems to be stable now, I can review this later if that's necessary.